requirements.txt
torch>=1.12.0
torchvision>=0.13.0
Pillow>=8.0.0
numpy>=1.21.0
onnx>=1.12.0
onnxruntime>=1.12.0
psutil>=5.8.0
tqdm>=4.64.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0 